Added Admin Panel and structured the files
