+++ .spyproject/config/backups/workspace.ini.bak
@@ -4,7 +4,7 @@ save_data_on_exit = True
 save_history = True
 save_non_project_files = False
 project_type = 'empty-project-type'
-recent_files = ['app.py', 'templates\\stop_attendance.html', 'templates\\faculty_dashboard.html', 'templates\\filtered_data.html', 'templates\\modify_attendance.html']
+recent_files = ['app.py']
 
 [main]
 version = 0.2.0
