+++ admin/templates/add_student_list.html
@@ -1,0 +1,92 @@
+{% extends 'base.html' %}
+
+{% block title %} Add Student List {% endblock %}
+
+{% block content %}
+<div class="container-fluid">
+  <div class="row">
+    <!-- Sidebar -->
+    <nav id="sidebar" class="col-md-3 col-lg-2 d-md-block bg-light sidebar">
+      <div class="sidebar-sticky">
+        <ul class="nav flex-column">
+          <li class="nav-item">
+            <a class="nav-link" href="/admin/time_table_manage">
+              <i class="fas fa-clock mr-2"></i> Time Table Management
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/admin/faculty_manage">
+              <i class="fas fa-user-tie mr-2"></i> Faculty Management
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/admin/student_manage">
+              <i class="fas fa-user-graduate mr-2"></i> Student Management
+            </a>
+          </li>
+        </ul>
+      </div>
+    </nav>
+
+    <!-- Main Content -->
+    <main role="main" class="col-md-9 ml-sm-auto col-lg-10 px-md-4">
+      <div class="d-flex justify-content-between align-items-center pt-3 pb-2 mb-3 border-bottom">
+        <h1 class="h2">Add Student List</h1>
+      </div>
+      <div class="row">
+        <div class="col-md-8 mx-auto">
+          <div class="card">
+            <div class="card-header">
+              <h5 class="card-title">Student Details</h5>
+            </div>
+            <div class="card-body">
+              <form action="{{ url_for('admin.add_student_list') }}" method="post">
+                <div class="form-group">
+                  <label for="academic_year">Academic Year:</label>
+                  <input type="text" class="form-control" id="academic_year" name="academic_year" required>
+                </div>
+                <div class="form-group">
+                  <label for="department">Department:</label>
+                  <input type="text" class="form-control" id="department" name="department" required>
+                </div>
+                <div class="form-group">
+                  <label for="sem">Semester:</label>
+                  <input type="text" class="form-control" id="sem" name="sem" required>
+                </div>
+                <div class="form-group">
+                  <label for="subject">Subject:</label>
+                  <input type="text" class="form-control" id="subject" name="subject" required>
+                </div>
+                <button type="submit" class="btn btn-primary">Submit</button>
+              </form>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div class="row mt-4">
+        <div class="col-md-8 mx-auto">
+          <div class="card">
+            <div class="card-header">
+              <h5 class="card-title">Student List</h5>
+            </div>
+            <div class="card-body">
+              <table class="table">
+                <thead>
+                  <tr>
+                    <th>Enrollment</th>
+                    <th>Name</th>
+                    <th>Email</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  <!-- Populate student list here -->
+                </tbody>
+              </table>
+            </div>
+          </div>
+        </div>
+      </div>
+    </main>
+  </div>
+</div>
+{% endblock %}
