+++ admin/templates/admin_login.html
@@ -1,0 +1,30 @@
+{% extends 'base.html' %}
+    
+{% block title %} Admin Login {% endblock %}
+
+{% block content %}
+    <div class="d-flex justify-content-center align-items-center vh-100">
+        <div class="card p-5">
+            <div class="card-header text-center">
+                <h1 class="card-title">Admin Login</h1>
+            </div>
+            <div class="card-body">
+                <form action="{{ url_for('admin.admin_login') }}" method="post">
+                    {% if msg %}
+                        <div class="alert alert-danger" role="alert">
+                            {{ msg }}
+                        </div>
+                    {% endif %}
+                    <div class="form-group">
+                        <input id="admin_email" name="admin_email" type="text" placeholder="Email ID" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="password" name="password" type="password" placeholder="Password" class="form-control" required />
+                    </div>
+                    <button type="submit" class="btn btn-success btn-block">Login</button>
+                </form>
+                <p class="text-center mt-3">Don't have an account? <a href="{{ url_for('admin.admin_register') }}" class="btn btn-primary">Register</a></p>
+            </div>
+        </div>
+    </div>
+{% endblock %}
