+++ admin/templates/modify_subject_list.html
@@ -1,0 +1,101 @@
+{% extends 'base.html' %}
+
+{% block title %} Modify Subject List {% endblock %}
+
+{% block content %}
+<div class="container-fluid">
+  <div class="row">
+    <!-- Sidebar -->
+    <nav id="sidebar" class="col-md-3 col-lg-2 d-md-block bg-light sidebar">
+      <div class="sidebar-sticky">
+        <ul class="nav flex-column">
+          <li class="nav-item">
+            <a class="nav-link" href="/admin/time_table_manage">
+              <i class="fas fa-clock mr-2"></i> Time Table Management
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/admin/faculty_manage">
+              <i class="fas fa-user-tie mr-2"></i> Faculty Management
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/admin/student_manage">
+              <i class="fas fa-user-graduate mr-2"></i> Student Management
+            </a>
+          </li>
+        </ul>
+      </div>
+    </nav>
+
+    <!-- Main Content -->
+    <main role="main" class="col-md-9 ml-sm-auto col-lg-10 px-md-4">
+      <div class="d-flex justify-content-between align-items-center pt-3 pb-2 mb-3 border-bottom">
+        <h1 class="h2">Modify Subject List</h1>
+      </div>
+      <div class="row mt-4">
+        <div class="col-md-8 mx-auto">
+          <div class="card">
+            <div class="card-header">
+              <h5 class="card-title">Modify Subject</h5>
+            </div>
+            <div class="card-body">
+              <form action="{{ url_for('admin.modify_subject_list') }}" method="post">
+                <div class="form-group">
+                  <label for="department">Department:</label>
+                  <input type="text" class="form-control" id="department" name="department" required>
+                </div>
+                <div class="form-group">
+                  <label for="semester">Semester:</label>
+                  <input type="text" class="form-control" id="semester" name="semester" required>
+                </div>
+                <div class="form-group">
+                  <label for="subject">Subject:</label>
+                  <!-- Dropdown menu of subjects -->
+                  <select class="form-control" id="subject" name="subject" required>
+                  
+                  </select>
+                </div>
+              </form>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div class="row mt-4">
+        <div class="col-md-8 mx-auto">
+          <div class="card">
+            <div class="card-header">
+              <h5 class="card-title">New Values</h5>
+            </div>
+            <div class="card-body">
+              <form action="{{ url_for('admin.modify_subject_list') }}" method="post">
+                <table class="table">
+                  <thead>
+                    <tr>
+                      <th>Subject Name</th>
+                      <th>Subject Code</th>
+                      <th>Lecture</th>
+                      <th>Lab</th>
+                      <th>Tutorial</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    <tr>
+                      <td><input type="text" class="form-control" name="subject_name" required></td>
+                      <td><input type="number" class="form-control" name="subject_code" min="0" required></td>
+                      <td class="text-center"><input type="checkbox" class="form-check-input" name="lecture"></td>
+                      <td class="text-center"><input type="checkbox" class="form-check-input" name="lab"></td>
+                      <td class="text-center"><input type="checkbox" class="form-check-input" name="tutorial"></td>
+                    </tr>
+                  </tbody>
+                </table>
+                <button type="submit" class="btn btn-primary">Submit New Values</button>
+              </form>
+            </div>
+          </div>
+        </div>
+      </div>
+    </main>
+  </div>
+</div>
+{% endblock %}
