+++ faculty/templates/faculty_register.html
@@ -1,0 +1,54 @@
+{% extends 'base.html' %}
+    
+{% block title %} Faculty Registration {% endblock %}
+
+{% block content %}
+    <div class="d-flex justify-content-center align-items-center vh-100">
+        <div class="card p-5">
+            <div class="card-header text-center">
+                <h1 class="card-title">Faculty Registration</h1>
+            </div>
+            <div class="card-body">
+                <form action="{{ url_for('faculty.faculty_register') }}" method="post">
+                    <div class="form-group">
+                        <input id="faculty_id" name="faculty_id" type="text" placeholder="Faculty ID" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="name" name="name" type="text" placeholder="Full Name" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="email" name="email" type="email" placeholder="Email" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="password" name="password" type="password" placeholder="Password" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="teaching_class" name="teaching_class" type="text" placeholder="Teaching classes (e.g., 3CE1_DS_202040301)" class="form-control" required />
+                    </div>
+
+                    <div id="class-container"></div>
+                    <button id="add-class" class="btn btn-info">Add Another Class</button><br><br>
+
+                    <button type="submit" class="btn btn-success btn-block">Register</button>
+                </form>
+                <p class="text-center mt-3">Already have an account? <a href="{{ url_for('faculty.faculty_login') }}">Login</a></p>
+            </div>
+        </div>
+    </div>
+
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <script>
+        $(document).ready(function () {
+            var maxClasses = 3; 
+            var classesCount = 0;
+
+            $("#add-class").click(function () {
+                if (classesCount < maxClasses) {
+                    var inputField = '<div class="form-group"><input type="text" name="teaching_class" placeholder="Teaching classes (e.g., 3CE1_DS_202040301)" class="form-control" required /></div>';
+                    $("#class-container").append(inputField);
+                    classesCount++;
+                }
+            });
+        });
+    </script>
+{% endblock %}
