+++ faculty/templates/modify_attendance.html
@@ -1,0 +1,50 @@
+{% extends 'base.html' %}
+
+{% block title %} Modify Attendance {% endblock %}
+
+{% block content %}
+  <div class="d-flex justify-content-center align-items-center vh-100">
+    <div class="card p-4 position-relative">
+      <div class="card-header text-center">
+        <h1 class="card-title">Modify Attendance</h1>
+      </div>
+
+      <div class="card-body">
+          {% if msg %}
+            <div class="alert alert-danger">{{ msg }}</div>
+          {% endif %}
+        <form action="{{ url_for('faculty.modify_attendance') }}" method="post">
+          <div class="form-group">
+            <label for="subject" class="font-weight-bold">Subject: </label>
+            <select name="subject" id="subject" class="form-control" required>
+              <option value="" disabled selected>Select a Subject</option>
+              {% for subject in subjects %}
+                <option value="{{ subject }}">{{ subject }}</option>
+              {% endfor %}
+            </select>
+          </div> 
+           
+          <div class="form-group">
+            <label for="date" class="font-weight-bold">Date: </label>
+            <input type="date" name="date" class="form-control" required />
+          </div>
+
+          <div class="form-group">
+            <label for="enrollment" class="font-weight-bold">Enrollment: </label>
+            <input type="text" name="enrollment" placeholder="Enrollment No." class="form-control" required/>
+          </div>
+
+          <div class="form-group">
+         
+            <label for="new_attendance" class="font-weight-bold">New Attendance: </label>
+            <input type="text" name="new_attendance" placeholder="A: Absent, P: Present, L: Leave" class="form-control" required/>
+          </div>
+          <div class="text-center mt-4">
+              <button type="submit" class="btn btn-success">Save Changes</button>
+              <a href="/faculty/faculty_dashboard" class="btn btn-success">Home</a>
+          </div>
+        </form>
+      </div>
+    </div>
+  </div>
+{% endblock %}
