+++ faculty/templates/stop_attendance.html
@@ -1,0 +1,42 @@
+{% extends 'base.html' %}
+
+{% block title %} Stop Attendance {% endblock %}
+
+{% block content %}
+    <div class="d-flex justify-content-center align-items-center vh-100">
+        <div class="card p-5">
+            <div class="card-header text-center">
+                <h1 class="card-title">Attendance Taken</h1>
+            </div>
+            <div class="card-body text-center">
+                <p>Number of attendant(s): {{students_count}}</p>
+            {% if students_count>0 %}
+                <p>Student(s) who enrolled in this attendance session of {{selected_subject}}:</p>
+                <table class="table">
+                    <thead>
+                        <tr>
+                            <th>Enrollment</th>
+                            <th>Name</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        {% for student in students_data %}
+                            <tr>
+                                <td>{{student[0]}}</td>
+                                <td>{{student[1]}}</td>
+                            </tr>
+                        {% endfor %}
+                    </tbody>
+                </table>
+            {% endif %}
+                <div class="text-center mt-4">
+                    <a href="/faculty/faculty_dashboard" class="btn btn-primary">Home</a>
+                    <a href="/faculty/logout" class="btn btn-primary">Logout</a>
+                    {% if students_count>0 %}
+                    <a href="/faculty/download_attendance" class="btn btn-primary">Download</a>
+                    {% endif %}
+                </div>
+            </div>
+        </div>
+    </div>
+{% endblock %}
