+++ student/templates/attendance_marked.html
@@ -1,0 +1,24 @@
+{% extends 'base.html' %}
+
+{% block title %} Attendance Marked {% endblock %}
+
+{% block content %}
+    <div class="d-flex justify-content-center align-items-center vh-100">
+        <div class="card p-5">
+            <div class="card-header text-center">
+                <h1 class="card-title">Attendance Marked</h1>
+            </div>
+            <div class="card-body text-center">
+                
+                <p>Your attendance has been marked for the following:</p>
+                <p>Subject: {{ selected_subject }}</p>
+                <p>Enrollment Number: {{ enrollment }}</p>
+
+                <div class="text-center mt-4">
+                    <a href="/student/student_dashboard" class="btn btn-primary">Home</a>
+                    <a href="/student/logout" class="btn btn-primary">Logout</a>
+                </div>
+            </div>
+        </div>
+    </div>
+{% endblock %}
