+++ student/templates/student_dashboard.html
@@ -1,0 +1,64 @@
+{% extends 'base.html' %}
+    
+{% block title %} Student Dashboard {% endblock %}
+
+{% block content %}
+    <div class="d-flex justify-content-center align-items-center vh-100">
+        <div class="card p-5">
+            <div class="card-header text-center">
+                <h1 class="card-title">Student Dashboard</h1>
+            </div>
+            <div class="card-body text-center">
+                {% if msg %}
+                    <div class="alert alert-danger">{{ msg }}</div>
+                {% endif %}
+                <form action="{{ url_for('student.student_dashboard') }}" method="post">
+                    <div class="form-group">
+                        <select name="subject" id="subject" class="form-control" required>
+                            <option value="" disabled selected>Select a Subject</option>
+                            {% for subject in subjects %}
+                                <option value="{{ subject[0] }}">{{ subject[0] }}</option>
+                            {% endfor %}
+                        </select>
+                        <div class="invalid-feedback">Please select a subject.</div>
+                    </div>
+
+                    <input type="text" name="enrollment" placeholder="Enrollment no" class="form-control" required/><br>
+                    <input type="text" name="attendance_id" id="attendance_id" placeholder="Attendance ID" class="form-control" required /><br>
+
+                    <button id="submit-button" type="submit" class="btn btn-success btn-block" disabled>Submit</button>
+                </form>
+
+                {% if timer_value is not none %}
+                    <div id="countdown">Time remaining: <div id="timer"></div></div>
+                {% endif %}
+            </div>
+        </div>
+    </div>
+
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.3.2/socket.io.js"></script>
+    <script>
+        var socket = io.connect('http://' + document.domain + ':' + location.port);
+
+        socket.on('connect', function() {
+            console.log('Connected to the server');
+        });
+
+        socket.on('update_timer', function(timerValue) {
+            document.getElementById('timer').innerText = timerValue;
+        });
+    </script>
+
+    <script>
+        const subjectSelect = document.getElementById('subject');
+        const submitButton = document.getElementById('submit-button');
+
+        subjectSelect.addEventListener('change', () => {
+            if (subjectSelect.value) {
+                submitButton.removeAttribute('disabled');
+            } else {
+                submitButton.setAttribute('disabled', 'disabled');
+            }
+        });
+    </script>
+{% endblock %}
