+++ student/templates/student_register.html
@@ -1,0 +1,32 @@
+{% extends 'base.html' %}
+    
+{% block title %} Student Registration {% endblock %}
+
+{% block content %}
+    <div class="d-flex justify-content-center align-items-center vh-100">
+        <div class="card p-5">
+            <div class="card-header text-center">
+                <h1 class="card-title">Student Registration</h1>
+            </div>
+            <div class="card-body">
+                <form action="{{ url_for('student.student_register') }}" method="post">
+                    {{ msg }}
+                    <div class="form-group">
+                        <input id="enrollment" name="enrollment" type="text" placeholder="Enrollment Number" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="name" name="name" type="text" placeholder="Full Name" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="student_class" name="student_class" type="text" placeholder="Studying Class (e.g., 5CE1)" class="form-control" required />
+                    </div>
+                    <div class="form-group">
+                        <input id="password" name="password" type="password" placeholder="Password" class="form-control" />
+                    </div>
+                    <button type="submit" class="btn btn-success btn-block">Register</button>
+                </form>
+                <p class="text-center mt-3">Already have an account? <a href="{{ url_for('student.student_login') }}">Login</a></p>
+            </div>
+        </div>
+    </div>
+{% endblock %}
